style: refactor ordered facets
