+++ src/components/sections/ProductGallery/useOrderedFacets.ts
@@ -1,0 +1,29 @@
+import type {
+  FacetedFilter_FacetsFragment,
+  ProductGalleryQueryQuery,
+} from '@generated/graphql'
+import { useMemo, useRef } from 'react'
+
+export const useOrderedFacets = (data?: ProductGalleryQueryQuery) => {
+  const facets = useRef<FacetedFilter_FacetsFragment[]>([])
+
+  return useMemo(() => {
+    if (data) {
+      const orderFacets = data.search.facets.map((facet) => {
+        if (facet.type === 'BOOLEAN') {
+          const sortedValues = facet.values.sort((a, b) =>
+            a.label.localeCompare(b.label)
+          )
+
+          return { ...facet, values: sortedValues }
+        }
+
+        return facet
+      })
+
+      facets.current = orderFacets
+    }
+
+    return facets.current
+  }, [data, facets])
+}
